Homework 5: CPU alarm
Add system call to set periodic alarm for the current process, the alarm
will call a user-defined function every X ticks.

LAPIC generates timer interrupts for every tick and if the alarm is set
up, and the tick counts to 0, the alarm function is called. Note that
the function is called in user space by setting return-to EIP to the
be function in interrupt handler. The original EIP is also pushed to
user stack so the alarm function will return to the original interrupt
position.

This serves as a simplified version of a process signal.