Homework 10: big files
Change inode format so that it has an additional doubly-indirect
list of list of blocks, increasing the max file size from 140 sectors to
16523 sectors. 16523 = 11+128+128*128. Number of direct blocks is
reduced by 1, to provide an entry for the doubly-indirect block table.

Abstract a function to get the block number for a given index in the
specified block. Can easily be used to create more levels of indirect
block tables.