Homework 11: xv6 log
Streamline commit() to avoid copy again from the log blocks, because
the target blocks are guaranteed to be still in cache and locked. This
is because if it's not true, commit(), when it's called, can't be sure
the target blocks have the content and can be copied into log blocks.

So only need to directly read it from cache and sync to disk in
install_trans().